refactor: consolidate window layout system into manager
Merges the logic from `window-layout.lua` (state persistence, window tracking, screen matching) directly into the newly scaffolded `window-manager.lua` to centralize all custom layout and screen event tracking capabilities. Deletes the obsolete `window-layout.lua` and updates `init.lua` accordingly.
